requirements.txt
fastapi>=0.120.2
litellm>=1.79.0
python-dotenv>=1.2.1
streamlit>=1.51.0
uvicorn>=0.38.0